PiGui: move widget declarations into a separate file
